requirements.txt

flaml[automl]>=2.3.3
html5lib>=1.1
langchain-community>=0.3.14
langchain-openai>=0.3.0
langchain>=0.3.14
lxml>=5.3.0
openai>=1.59.6
pandas>=2.2.3
plotly>=5.24.1
pydantic-ai-slim>=0.7.5
quandl>=3.7.0
streamlit>=1.41.1
tavily-python>=0.5.0
tqdm>=4.67.1
trafilatura>=2.0.0
yfinance>=0.2.51